Modal Footer rewritten to TS
